refactor(zero-cache): move invalidation table stmts to separate file
